+++ app/Http/Controllers/BooklayoutsController.php
@@ -95,6 +95,11 @@ class BooklayoutsController extends Controller
     public function destroy($id)
     {
         $Booklayout = Booklayout::findOrFail($id);
+        $BooksCount = $Booklayout->book()->count();
+
+        if ($BooksCount > 0) {
+            return response()->json(['message' => 'Cannot delete with associated books.'], 422);
+        }
         $Booklayout->delete();
         return  response(null,204);
     }
